+++ pynfse/nfse/pr/curitiba/_dsig.py
