+++ pynfse/nfse/pr/curitiba/_schema.py
