+++ pynfse/nfse/pr/curitiba/serializacao.py
@@ -1,6 +1,5 @@
+from pynfse.nfse.pr.curitiba import _schema as nfse_schema
 from pyxb import BIND
-
-from PyNFSe.nfse.pr.curitiba import _schema as nfse_schema
 
 
 def consulta_nfse_por_numero(prestador, numero_nfse):
