+++ tests/pr/curitiba/test_curitiba_serializacao_comunicacao.py
@@ -1,7 +1,8 @@
 import datetime
 import unittest
 
-import PyNFSe.nfse.pr.curitiba.serializacao as s
+import pynfse.nfse.pr.curitiba.serializacao as s
+
 from tests.pr.curitiba.basetestesserializacao import BaseTestesSerializacao, xml_expected
